refactor: remove game_renderer.py and consolidate with ui_renderer.py
- Removed game_renderer.py as it was a duplicate of ui_renderer.py.
- Moved the draw_flies() function from game_logic.py to ui_renderer.py for better organization.
